pyproject.toml
[tool.poetry]
name = "netmikro"
version = "0.1.0"
description = "Netmikro is a simple library that provides an easy way to manage Mikrotik routers, simplifying configuration and monitoring tasks."
authors = ["Henrique Sebastião <contato@henriquesebastiao.com>"]
readme = "README.md"
packages = [{include = "netmikro"}]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Telecommunications Industry",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Networking :: Firewalls",
    "Topic :: System :: Networking",
    "Topic :: System :: Networking :: Monitoring :: Hardware Watchdog",
    "Topic :: Utilities",
]

[tool.poetry.urls]
"Documentation" = "https://netmikro.henriquesebastiao.com"
"Homepage" = "https://netmikro.henriquesebastiao.com"
"Repository" = "https://github.com/henriquesebastiao/netmikro"
"Bug Tracker" = "https://github.com/henriquesebastiao/netmikro/issues"

[tool.poetry.dependencies]
python = "^3.11.9"
netmiko = "^4.2.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-cov = "^4.1.0"
blue = "^0.9.1"
isort = "^5.12.0"
ruff = "^0.0.287"
taskipy = "^1.12.0"

[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.2.8"
mkdocstrings = "^0.23.0"
mkdocstrings-python = "^1.6.2"
mkdocs-macros-plugin = "^1.0.5"
jinja2 = "^3.1.3"
pygments = "^2.17.2"
pymdown-extensions = "^10.7"
mkdocs-git-committers-plugin-2 = "^2.2.3"
mkdocs-git-authors-plugin = "^0.7.2"
mkdocs-git-revision-date-localized-plugin = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
python_files = "test.py tests.py test_*.py tests_*.py *_test.py *_tests.py"

[tool.coverage.run]
omit = ["_utils.py", "hand.py"]

[tool.isort]
profile = "black"
line_length = 79

[tool.ruff]
line-length = 79
exclude = ["netmikro/__init__.py"]
tab-size = 4

[tool.ruff.per-file-ignores]
"tests/*" = ["F401", "F811"]
"__init__.py" = ["F401"]

[tool.taskipy.tasks]
lint = "ruff . && blue --check --diff . && isort --check --diff ."
format = 'blue .  && isort .'
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=netmikro -vv"
post_test = "coverage html"
export-requirements = "rm requirements.txt && poetry export -f requirements.txt --output requirements.txt --without-hashes"
